pyproject.toml
[project]
name = "ykicks"
version = "0.1.0"
description = "Real-time monitoring of Yearn’s new generation (NG) dumping services"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ruff==0.11.5",
    "mypy==1.15.0",
    "python-dotenv==1.1.0",
    "python-telegram-bot==22.0",
    "eth-ape==0.8.36",
    "silverback==0.7.27",
    "ape-etherscan==0.8.4",
]

[tool.ruff]
line-length = 120
target-version = "py310"
lint.select = ["E", "F", "I", "W"]  # pycodestyle, pyflakes, isort, and warnings
fix = true
exclude = ["__pypackages__"]
lint.ignore = ['E501']

[tool.ruff.format]
quote-style = "double"
docstring-code-format = true

[tool.mypy]
mypy_path = "src"
python_version = "3.12"
ignore_missing_imports = true
strict = true
check_untyped_defs = true
disallow_untyped_defs = true
warn_unused_ignores = true
disable_error_code = ["misc"]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"